requirements/base.txt
#Django packages
django==3.1
djangorestframework==3.11.1
django-filter==2.3.0
django-bootstrap4==2.2.0
django-crispy-forms==1.9.2
django-tinymce==3.0.2
django-environ==0.4.5

# Tools
mysqlclient==2.0.1
celery==4.4.7
flower==0.9.5
Pillow==7.2.0
requests==2.24.0
spotipy==2.14.0
PyJWT==1.7.1
argon2-cffi==20.1.0
